refactor(basemap): remove leading underscores from module names
Rename _basemap.py -> basemap.py, _tiles.py -> tiles.py,
_warp.py -> warp.py to match the project convention (no other
module in pyramids uses leading underscores). Update all imports
in source and tests.
