+++ src/pyramids/dataset/ops/analysis.py
@@ -953,7 +953,7 @@ class Analysis:
         cleo.plot(**kwargs)
 
         if basemap is not None:
-            from pyramids.basemap._basemap import add_basemap
+            from pyramids.basemap.basemap import add_basemap
 
             source = basemap if isinstance(basemap, str) else None
             add_basemap(cleo.ax, crs=self.epsg, source=source)
