+++ src/pyramids/netcdf/ugrid/dataset.py
@@ -684,7 +684,7 @@ class UgridDataset:
         )
 
         if basemap is not None:
-            from pyramids.basemap._basemap import add_basemap
+            from pyramids.basemap.basemap import add_basemap
 
             source = basemap if isinstance(basemap, str) else None
             add_basemap(result, crs=self.epsg, source=source)
